+++ lib/dial_codes.dart
@@ -3,3 +3,5 @@ library;
 
 export 'src/models/country.dart';
 export 'src/dial_codes_service.dart';
+export 'src/widgets/country_picker_dialog.dart';
+export 'src/widgets/country_dropdown.dart';